shift 'issue search' code to projectrelated.js for extendibility
